# services/web-bff/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: HooshiX Web BFF Public API
  version: 2.0.0
  description: Canonical browser-facing REST contract owned by Web BFF. All examples use synthetic .test identities and non-production placeholder credentials.
servers:
  - url: /
security:
  - BrowserSession: []
paths:
...
  /api/v1/identity/tenants:
    get:
      operationId: listIdentityTenants
      summary: List tenants available to the authenticated user
      security: [{ BrowserSession: [] }]
      x-hooshix-consumer-example: { method: GET, path: /api/v1/identity/tenants }
      responses:
        '200': { description: Tenant choices., content: { application/json: { schema: { $ref: '#/components/schemas/TenantListResponse' } } } }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
    post:
      operationId: createIdentityTenant
      summary: Create a tenant for the authenticated user
      security: [{ BrowserSession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CsrfToken' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateTenantRequest' } } } }
      x-hooshix-consumer-example: { method: POST, path: /api/v1/identity/tenants, body: { name: Sample Tenant, slug: sample-tenant } }
      responses:
        '200': { description: Created tenant., content: { application/json: { schema: { $ref: '#/components/schemas/TenantCreatedResponse' } } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/RegistrationRejected' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /api/v1/identity/tenant-selection:
    post:
      operationId: selectIdentityTenant
      summary: Select an active tenant membership
      security: [{ BrowserSession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CsrfToken' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/SelectTenantRequest' } } } }
      x-hooshix-consumer-example: { method: POST, path: /api/v1/identity/tenant-selection, body: { membershipId: 11111111-1111-4111-8111-111111111111 } }
      responses:
        '200': { description: Tenant selected and CSRF state rotated., content: { application/json: { schema: { $ref: '#/components/schemas/TenantSelectionResponse' } } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/RegistrationRejected' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /api/v1/identity/invitations:
    get:
      operationId: listIdentityTenantInvitations
      summary: List invitations for the selected tenant
      security: [{ BrowserSession: [] }]
      x-hooshix-consumer-example: { method: GET, path: /api/v1/identity/invitations }
      responses:
        '200': { description: Tenant invitations., content: { application/json: { schema: { $ref: '#/components/schemas/InvitationListResponse' } } } }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
    post:
      operationId: inviteIdentityTenantMember
      summary: Invite an existing contact to the selected tenant
      security: [{ BrowserSession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CsrfToken' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/InviteRequest' } } } }
      x-hooshix-consumer-example: { method: POST, path: /api/v1/identity/invitations, body: { targetContactId: 11111111-1111-4111-8111-111111111111 } }
      responses:
        '200': { description: Invitation created., content: { application/json: { schema: { $ref: '#/components/schemas/InvitationCreatedResponse' } } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/RegistrationRejected' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /api/v1/identity/invitations/received:
    get:
      operationId: listReceivedIdentityTenantInvitations
      summary: List invitations received by the authenticated user
      security: [{ BrowserSession: [] }]
      x-hooshix-consumer-example: { method: GET, path: /api/v1/identity/invitations/received }
      responses:
        '200': { description: Received invitations., content: { application/json: { schema: { $ref: '#/components/schemas/InvitationListResponse' } } } }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /api/v1/identity/invitations/{invitationId}/accept:
    post:
      operationId: acceptIdentityTenantInvitation
      summary: Accept a tenant invitation
      security: [{ BrowserSession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CsrfToken' }, { $ref: '#/components/parameters/InvitationId' }]
      x-hooshix-consumer-example: { method: POST, path: /api/v1/identity/invitations/11111111-1111-4111-8111-111111111111/accept }
      responses:
        '200': { description: Invitation accepted., content: { application/json: { schema: { $ref: '#/components/schemas/AcceptedInvitationResponse' } } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/RegistrationRejected' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /api/v1/identity/memberships/{membershipId}:
    delete:
      operationId: removeIdentityTenantMembership
      summary: Remove a membership from the selected tenant
      security: [{ BrowserSession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CsrfToken' }, { $ref: '#/components/parameters/MembershipId' }]
      x-hooshix-consumer-example: { method: DELETE, path: /api/v1/identity/memberships/11111111-1111-4111-8111-111111111111 }
      responses:
        '200': { description: Removal result., content: { application/json: { schema: { $ref: '#/components/schemas/RemovalResultResponse' } } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/RegistrationRejected' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /api/v1/identity/tenants/{tenantId}/suspend:
    post:
      operationId: suspendIdentityTenant
      summary: Request platform-authorized tenant suspension
      security: [{ BrowserSession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CsrfToken' }, { $ref: '#/components/parameters/TenantId' }]
      x-hooshix-consumer-example: { method: POST, path: /api/v1/identity/tenants/11111111-1111-4111-8111-111111111111/suspend }
      responses:
        '200': { description: Suspension accepted for durable reconciliation., content: { application/json: { schema: { $ref: '#/components/schemas/TenantLifecycleResultResponse' } } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/RegistrationRejected' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /api/v1/identity/tenants/{tenantId}/resume:
    post:
      operationId: resumeIdentityTenant
      summary: Request platform-authorized tenant resumption
      security: [{ BrowserSession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CsrfToken' }, { $ref: '#/components/parameters/TenantId' }]
      x-hooshix-consumer-example: { method: POST, path: /api/v1/identity/tenants/11111111-1111-4111-8111-111111111111/resume }
      responses:
        '200': { description: Resumption accepted for durable reconciliation., content: { application/json: { schema: { $ref: '#/components/schemas/TenantLifecycleResultResponse' } } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/RegistrationRejected' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /api/v1/identity/tenants/{tenantId}:
    delete:
      operationId: deleteIdentityTenant
      summary: Request owner-authorized tenant deletion
      description: The tenant enters durable cleanup; the browser session is rotated to onboarding mode and DELETED is reported only after Authorization acknowledges cleanup.
      security: [{ BrowserSession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CsrfToken' }, { $ref: '#/components/parameters/TenantId' }]
      x-hooshix-consumer-example: { method: DELETE, path: /api/v1/identity/tenants/11111111-1111-4111-8111-111111111111 }
      responses:
        '200': { description: Deletion accepted and browser security state rotated., content: { application/json: { schema: { $ref: '#/components/schemas/TenantLifecycleResultResponse' } } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/RegistrationRejected' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /api/v1/identity/tenants/{tenantId}/restore:
    post:
      operationId: restoreIdentityTenant
      summary: Request platform-authorized restore before purge starts
      security: [{ BrowserSession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CsrfToken' }, { $ref: '#/components/parameters/TenantId' }]
      x-hooshix-consumer-example: { method: POST, path: /api/v1/identity/tenants/11111111-1111-4111-8111-111111111111/restore }
      responses:
        '200': { description: Restore accepted for durable reconciliation., content: { application/json: { schema: { $ref: '#/components/schemas/TenantLifecycleResultResponse' } } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/RegistrationRejected' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /api/v1/identity/invitations/{invitationId}/decline:
    post:
      operationId: declineIdentityTenantInvitation
      summary: Decline a received pending invitation
      security: [{ BrowserSession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CsrfToken' }, { $ref: '#/components/parameters/InvitationId' }]
      x-hooshix-consumer-example: { method: POST, path: /api/v1/identity/invitations/11111111-1111-4111-8111-111111111111/decline }
      responses:
        '200': { description: Invitation declined., content: { application/json: { schema: { $ref: '#/components/schemas/InvitationStateResponse' } } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/RegistrationRejected' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /api/v1/identity/invitations/{invitationId}/revoke:
    post:
      operationId: revokeIdentityTenantInvitation
      summary: Revoke a pending invitation in the selected tenant
      security: [{ BrowserSession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CsrfToken' }, { $ref: '#/components/parameters/InvitationId' }]
      x-hooshix-consumer-example: { method: POST, path: /api/v1/identity/invitations/11111111-1111-4111-8111-111111111111/revoke }
      responses:
        '200': { description: Invitation revoked., content: { application/json: { schema: { $ref: '#/components/schemas/InvitationStateResponse' } } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/RegistrationRejected' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /api/v1/identity/invitations/{invitationId}/reissue:
    post:
      operationId: reissueIdentityTenantInvitation
      summary: Reissue a terminal invitation with a fresh seven-day expiry
      security: [{ BrowserSession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CsrfToken' }, { $ref: '#/components/parameters/InvitationId' }]
      x-hooshix-consumer-example: { method: POST, path: /api/v1/identity/invitations/11111111-1111-4111-8111-111111111111/reissue }
      responses:
        '200': { description: Invitation reissued., content: { application/json: { schema: { $ref: '#/components/schemas/InvitationCreatedResponse' } } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/RegistrationRejected' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
...
components:
  securitySchemes:
    BrowserSession:
      type: apiKey
      in: cookie
      name: __Host-sajtech-session
      description: HttpOnly same-origin session. Its value is never exposed in this contract or browser storage.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Caller-generated UUIDv4 idempotency identity. Replays use the same value.
      schema:
        type: string
        format: uuid
        pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Current session-bound synchronizer token returned by a reviewed BFF response; do not persist it.
      schema: { type: string, minLength: 32, maxLength: 512 }
...
    MembershipId: { name: membershipId, in: path, required: true, schema: { $ref: '#/components/schemas/UuidV4' } }
    TenantId: { name: tenantId, in: path, required: true, schema: { $ref: '#/components/schemas/UuidV4' } }
...
    InvitationId: { name: invitationId, in: path, required: true, schema: { $ref: '#/components/schemas/UuidV4' } }
...
  schemas:
...
    UuidV4:
      type: string
      format: uuid
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
      example: 11111111-1111-4111-8111-111111111111
...
    CreateTenantRequest:
      type: object
      additionalProperties: false
      required: [name, slug]
      example: { name: Sample Tenant, slug: sample-tenant }
      properties:
        name: { type: string, minLength: 1, maxLength: 120, example: Sample Tenant }
        slug:
          type: string
          minLength: 3
          maxLength: 63
          pattern: '^[a-z0-9](?:[a-z0-9-]{1,61}[a-z0-9])$'
          example: sample-tenant
    SelectTenantRequest:
      type: object
      additionalProperties: false
      required: [membershipId]
      example: { membershipId: 11111111-1111-4111-8111-111111111111 }
      properties: { membershipId: { $ref: '#/components/schemas/UuidV4' } }
    InviteRequest:
      type: object
      additionalProperties: false
      required: [targetContactId]
      example: { targetContactId: 11111111-1111-4111-8111-111111111111 }
      properties: { targetContactId: { $ref: '#/components/schemas/UuidV4' } }
...
    TenantChoice:
      type: object
      additionalProperties: false
      required: [tenantId, membershipId, name, slug]
      example: { tenantId: 11111111-1111-4111-8111-111111111111, membershipId: 22222222-2222-4222-8222-222222222222, name: Sample Tenant, slug: sample-tenant }
      properties:
        tenantId: { $ref: '#/components/schemas/UuidV4' }
        membershipId: { $ref: '#/components/schemas/UuidV4' }
        name: { type: string, maxLength: 120, example: Sample Tenant }
        slug: { type: string, maxLength: 63, example: sample-tenant }
    TenantListResponse:
      type: object
      additionalProperties: false
      required: [tenants]
      example: { tenants: [{ tenantId: 11111111-1111-4111-8111-111111111111, membershipId: 22222222-2222-4222-8222-222222222222, name: Sample Tenant, slug: sample-tenant }], suggestedMembershipId: 22222222-2222-4222-8222-222222222222 }
      properties:
        tenants: { type: array, items: { $ref: '#/components/schemas/TenantChoice' } }
        suggestedMembershipId: { anyOf: [{ $ref: '#/components/schemas/UuidV4' }, { type: 'null' }] }
    TenantCreatedResponse:
      type: object
      additionalProperties: false
      required: [tenantId, membershipId, lifecycle]
      example: { tenantId: 11111111-1111-4111-8111-111111111111, membershipId: 22222222-2222-4222-8222-222222222222, lifecycle: ACTIVE }
      properties:
        tenantId: { $ref: '#/components/schemas/UuidV4' }
        membershipId: { $ref: '#/components/schemas/UuidV4' }
        lifecycle: { type: string, example: ACTIVE }
    TenantSelectionResponse:
      type: object
      additionalProperties: false
      required: [csrfToken, tenantId, membershipId, mode]
      example: { csrfToken: synthetic-csrf-token-not-a-secret, tenantId: 11111111-1111-4111-8111-111111111111, membershipId: 22222222-2222-4222-8222-222222222222, mode: TENANT_AUTHENTICATED }
      properties:
        csrfToken: { type: string, minLength: 32, maxLength: 512, example: synthetic-csrf-token-not-a-secret }
        tenantId: { $ref: '#/components/schemas/UuidV4' }
        membershipId: { $ref: '#/components/schemas/UuidV4' }
        mode: { type: string, enum: [TENANT_AUTHENTICATED], example: TENANT_AUTHENTICATED }
    InvitationCreatedResponse:
      type: object
      additionalProperties: false
      required: [invitationId, expiresAt]
      example: { invitationId: 11111111-1111-4111-8111-111111111111, expiresAt: '2030-01-01T00:00:00Z' }
      properties:
        invitationId: { $ref: '#/components/schemas/UuidV4' }
        expiresAt: { type: string, format: date-time, example: '2030-01-01T00:00:00Z' }
    TenantLifecycleResultResponse:
      type: object
      additionalProperties: false
      required: [tenantId, lifecycle, targetLifecycle, pending]
      example: { tenantId: 11111111-1111-4111-8111-111111111111, lifecycle: ACTIVE, targetLifecycle: SUSPENDED, pending: true }
      properties:
        tenantId: { $ref: '#/components/schemas/UuidV4' }
        lifecycle: { type: string, enum: [PROVISIONING, ACTIVE, SUSPENDED, DELETING, DELETED] }
        targetLifecycle: { type: string, enum: [PROVISIONING, ACTIVE, SUSPENDED, DELETING, DELETED] }
        pending: { type: boolean }
        csrfToken:
          anyOf:
            - { type: string, minLength: 32, maxLength: 512 }
            - { type: 'null' }
        mode:
          anyOf:
            - { type: string, enum: [AUTHENTICATED_ONBOARDING, TENANT_AUTHENTICATED] }
            - { type: 'null' }
...
    InvitationListResponse:
      type: object
      additionalProperties: false
      required: [invitations]
      example: { invitations: [{ invitationId: 11111111-1111-4111-8111-111111111111, tenantId: 22222222-2222-4222-8222-222222222222, tenantName: Sample Tenant, tenantSlug: sample-tenant, state: PENDING, expiresAt: '2030-01-01T00:00:00Z' }] }
      properties:
        invitations: { type: array, maxItems: 200, items: { $ref: '#/components/schemas/InvitationSummaryResponse' } }
    InvitationStateResponse:
      type: object
      additionalProperties: false
      required: [invitationId, state]
      example: { invitationId: 11111111-1111-4111-8111-111111111111, state: DECLINED }
      properties:
        invitationId: { $ref: '#/components/schemas/UuidV4' }
        state: { type: string, enum: [DECLINED, REVOKED] }
    AcceptedInvitationResponse:
      type: object
      additionalProperties: false
      required: [tenantId, membershipId]
      example: { tenantId: 11111111-1111-4111-8111-111111111111, membershipId: 22222222-2222-4222-8222-222222222222 }
      properties:
        tenantId: { $ref: '#/components/schemas/UuidV4' }
        membershipId: { $ref: '#/components/schemas/UuidV4' }
    RemovalResultResponse:
      type: object
      additionalProperties: false
      required: [accepted, mode]
      example: { accepted: true, csrfToken: synthetic-csrf-token-not-a-secret, mode: AUTHENTICATED_ONBOARDING }
      properties:
        accepted: { type: boolean, example: true }
        csrfToken:
          anyOf:
            - { type: string, minLength: 32, maxLength: 512 }
            - { type: 'null' }
          example: synthetic-csrf-token-not-a-secret
        mode: { type: string, enum: [AUTHENTICATED_ONBOARDING, TENANT_AUTHENTICATED], example: AUTHENTICATED_ONBOARDING }
...
  responses:
    InvalidRequest:
      description: Request validation failed.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    InvalidSession:
      description: A supplied browser session is invalid.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Browser-origin or session-bound request protection rejected the request.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    RegistrationRejected:
      description: Registration precondition was not satisfied. Response is non-enumerating.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    DependencyUnavailable:
      description: The BFF runtime or an authoritative downstream dependency is unavailable.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...